moved d3locales to client
