+++ lib/client/chart.js
@@ -2,7 +2,7 @@
 
 var _ = require('lodash');
 var times = require('../times');
-var d3locales = require('../d3locales');
+var d3locales = require('./d3locales');
 var padding = { bottom: 30 };
 
 function init (client, d3, $) {
