+++ lib/client/d3locales.js
